TJBUFSIZE(): Fix int overflow w/ large dimensions
This is the same fix from 3de2da4d7c8c7b0f67c7330195d496f6689b7371
applied to the legacy TJBUFSIZE() function in the TurboJPEG 1.0 API.

+++ turbojpeg.c
@@ -577,7 +577,8 @@ DLLEXPORT unsigned long TJBUFSIZE(int width, int height)
   /* This allows for rare corner cases in which a JPEG image can actually be
      larger than the uncompressed input (we wouldn't mention it if it hadn't
      happened before.) */
-  retval = PAD(width, 16) * PAD(height, 16) * 6ULL + 2048ULL;
+  retval = (unsigned long long)PAD(width, 16) * PAD(height, 16) * 6ULL +
+           2048ULL;
 #if ULLONG_MAX > ULONG_MAX
   if (retval > (unsigned long long)((unsigned long)-1))
     THROWG("TJBUFSIZE(): Image is too large");
